+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.9.18
 * Add AHT20 temperature and humidity sensor
+* Add AT24C128 256 KB EEPROM
 * Add SSD1306 OLED
 * Add Dart-Java-Bridge example to generate graphics for the SSD1306 OLED 
 * Update c-periphery to v2.4.3